backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
pydantic==2.7.0
pydantic-settings==2.2.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.35
python-multipart==0.0.9
weasyprint>=63.0,<64.0
pypdf2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.0
openai==1.1.1
requests==2.31.0
httpx==0.27.0
textstat==0.7.3
beautifulsoup4==4.12.3
lxml==5.1.0
# Industry-standard resume parsing libraries
PyMuPDF==1.23.8
docx2txt==0.8
pdf2image>=1.16.0
Pillow>=10.0.0
# ATS Checker dependencies - TF-IDF enabled for industry-standard scoring
scikit-learn==1.3.2
# spacy==3.6.1  # Optional: Advanced NLP features (has fallback if not installed)
nltk==3.8.1
# Optional dependencies with fallback support
firebase-admin==6.5.0
stripe==10.12.0

